RAG/healthcare-evidence-rag/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
python-dotenv==1.0.1

pypdf==5.1.0
numpy==2.1.3
scikit-learn==1.5.2

sentence-transformers==3.3.1
faiss-cpu==1.9.0.post1
rank-bm25==0.2.2

requests==2.32.3
mlflow==2.19.0